various
- aid can be long
- add config param for mdns service name
- http host uses mdns service name instead of ip address
- decouple and flush socket after pairing before secure session
- disable nagle to force frame flushing
- disable event subscribe if no evented channels
- update read me
- strip to sparse required http headers

+++ bundles/org.openhab.binding.homekit/src/main/java/org/openhab/binding/homekit/internal/dto/Accessories.java
@@ -26,7 +26,7 @@ import org.eclipse.jdt.annotation.Nullable;
 public class Accessories {
     public List<Accessory> accessories;
 
-    public @Nullable Accessory getAccessory(Integer aid) {
+    public @Nullable Accessory getAccessory(Long aid) {
         return accessories.stream().filter(a -> aid.equals(a.aid)).findFirst().orElse(null);
     }
 }
